Remove duplicate toolbox actions for UML
