+++ gaphor/diagram/support.py
@@ -1,16 +1,15 @@
 """
-For ease of creation, maintain a mapping from UML Element to Diagram Item.
+For ease of creation, maintain a mapping from (UML) Element to Diagram Item.
 """
 
 from typing import Dict
 
-from gaphor import UML
-from gaphor.core.modeling import Presentation
+from gaphor.core.modeling import Element, Presentation
 
 
 def represents(uml_element):
     """
-    A decorator to assign a default UML Element type to a diagram item.
+    A decorator to assign a default (UML) Element type to a diagram item.
     """
 
     def wrapper(presentation):
@@ -21,7 +20,7 @@ def represents(uml_element):
 
 
 # Map UML elements to their (default) representation.
-_uml_to_item_map: Dict[UML.Element, Presentation] = {}
+_uml_to_item_map: Dict[Element, Presentation] = {}
 
 
 def get_diagram_item(element):
